feat: plan

+++ src/lib/component/remove.ts
@@ -2,6 +2,7 @@
 /* eslint-disable no-await-in-loop */
 
 import * as core from '@serverless-devs/core';
+import Table from 'tty-table';
 import { ICredentials } from '../interface/profile';
 import Client from '../client';
 import logger from '../../common/logger';
@@ -11,9 +12,9 @@ import Alias from './alias';
 import Version from './version';
 import * as HELP from '../help/remove';
 import _ from 'lodash';
-import { getCredentials } from '../utils';
+import { getCredentials, promptForConfirmOrDetails } from '../utils';
 
-const COMMAND: string[] = [
+export const COMMAND: string[] = [
   'service',
   'function',
   'trigger',
@@ -170,7 +171,6 @@ export default class Remove {
   async remove({ props, subCommand }: IRemove, inputs) {
     const {
       region,
-      assumeYes,
       onlyLocal,
       serviceName,
       functionName,
@@ -178,10 +178,28 @@ export default class Remove {
       versionId,
       aliasName,
     } = props;
+    let { assumeYes } = props;
+    let planArgs = '';
+
+    if (assumeYes !== true) {
+      try {
+        const removeStatus = await this.removePlan(subCommand, inputs);
+        if (removeStatus === 'quit') {
+          return;
+        }
+        if (removeStatus === 'assumeYes') {
+          assumeYes = true;
+          planArgs += '--assume-yes';
+        }
+      } catch (ex) {
+        // 异常：不作处理兜底
+        logger.debug(`error: ${ex.message}`);
+      }
+    }
 
     if (subCommand === 'layer') {
       const componentName = 'devsapp/fc-layer';
-      const componentInputs = this.genInputs(inputs, componentName, props);
+      const componentInputs = this.genInputs(inputs, componentName, props, planArgs);
       if (versionId) {
         return (await core.loadComponent(componentName)).deleteVersion(componentInputs);
       }
@@ -190,7 +208,7 @@ export default class Remove {
 
     if (subCommand === 'domain') {
       const componentName = 'devsapp/fc-deploy';
-      const componentInputs = this.genInputs(inputs, componentName, inputs.props);
+      const componentInputs = this.genInputs(inputs, componentName, inputs.props, planArgs);
       return (await core.loadComponent(componentName)).remove(componentInputs);
     }
 
@@ -222,8 +240,32 @@ export default class Remove {
     }
 
     const componentName = 'devsapp/fc-deploy';
-    const componentInputs = this.genInputs(inputs, componentName, inputs.props);
+    const componentInputs = this.genInputs(inputs, componentName, inputs.props, planArgs);
     return (await core.loadComponent(componentName)).remove(componentInputs);
+  }
+
+  // 没有子资源：能够被删除，不作处理。 返回undefined
+  // 存在子资源选择 no：一定不能被删除，需要退出程序。 返回quit
+  // 存在子资源选择 yes：需要强制删除所有资源，需要向下传递 assumeYes。  返回assumeYes
+  private async removePlan(subCommand, inputs) {
+    const planArgs = `--plan-type remove --sub-command ${subCommand || 'service'}`;
+    const planName = 'devsapp/fc-plan';
+    const planInputs = this.genInputs(_.cloneDeep(inputs), planName, inputs.props, planArgs);
+    const plan = await (await core.loadComponent(planName)).plan(planInputs);
+    logger.debug(JSON.stringify(plan));
+
+    if (!_.isEmpty(plan)) {
+      for (const planItem of plan) {
+        logger.debug(JSON.stringify(planItem));
+        if (_.isEmpty(planItem.data)) {
+          continue;
+        }
+        logger.log(planItem.title || '');
+        console.log(Table(planItem.header, planItem.data).render());
+      }
+      const assumeYes = await promptForConfirmOrDetails('Are you sure you want to delete these resources?');
+      return assumeYes ? 'assumeYes' : 'quit';
+    }
   }
 
   private genInputs({
@@ -232,7 +274,7 @@ export default class Remove {
     access,
     args,
     curPath,
-  }, componentName, props) {
+  }, componentName, props, appendArgs?) {
     return {
       project: {
         component: componentName,
@@ -241,7 +283,7 @@ export default class Remove {
       },
       appName,
       props,
-      args,
+      args: `${args || ''} ${appendArgs || ''}`,
       path: curPath,
     };
   }
